handle scan and pay callbacks via celery
